+++ framework/js/setup.js
@@ -15,4 +15,9 @@
       'overflowtouch',
       window.getComputedStyle && window.getComputedStyle(elem).getPropertyValue('-webkit-overflow-scrolling') == 'touch');
   });
+
+  // confirm to use jqLite
+  if (window.jQuery && angular.element === window.jQuery) {
+    console.warn('Onsen UI require jqLite. Load jQuery after loading AngularJS to fix this error. jQuery may break Onsen UI behavior.');
+  }
 })();